pls
Former-commit-id: ee1cc4abdd19c09f10d51008e2062287c4257773

+++ src/gamestatus/ReadableGameStatus.java
@@ -2,7 +2,7 @@ package gamestatus;
 
 public interface ReadableGameStatus {
 	double getExp();
-	double getMoney();
+	int getMoney();
 	int getLevel();
 	int getLives();
 	int getEnemiesLeft();
